requirements-comet.txt
absl-py
asttokens
astunparse
certifi
charset-normalizer
compressai
decorator
easydict
executing
filelock
flatbuffers
fsspec
gast
google-pasta
grpcio
h5py
idna
imageio
ipython
jedi
Jinja2
kagglehub
keras
lazy_loader
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
ml-dtypes
mpmath
namex
networkx
neptune
numpy
nibabel
nilearn
opencv-python
opt_einsum
optree
packaging
parso
pexpect
pillow
prompt_toolkit
protobuf
ptyprocess
pure_eval
Pygments
requests
rich
scikit-image
scipy
seaborn
setuptools
six
stack-data
sympy
tensorboard
tensorboard-data-server
tensorflow
termcolor
tifffile
torch==2.6.0
torchvision
traitlets
typing_extensions
urllib3
wcwidth
Werkzeug
wheel
wrapt
omegaconf
dotenv
piqa